+++ adaptive_navigation/example/README.md
@@ -1,16 +1,3 @@
-# example
+# Adaptive navigation sample app
 
-A new Flutter application.
-
-## Getting Started
-
-This project is a starting point for a Flutter application.
-
-A few resources to get you started if this is your first Flutter project:
-
-- [Lab: Write your first Flutter app](https://flutter.dev/docs/get-started/codelab)
-- [Cookbook: Useful Flutter samples](https://flutter.dev/docs/cookbook)
-
-For help getting started with Flutter, view our
-[online documentation](https://flutter.dev/docs), which offers tutorials,
-samples, guidance on mobile development, and a full API reference.
+This example application demonstrates how to use [`adaptive_navigation`](https://pub.dev/packages/adaptive_navigation) within a simple Flutter app.
