+++ adaptive_navigation/example/lib/custom_scaffold.dart
@@ -1,4 +1,4 @@
-import 'package:adaptive_navigation/scaffold.dart';
+import 'package:adaptive_navigation/adaptive_navigation.dart';
 import 'package:flutter/material.dart';
 
 class CustomScaffoldDemo extends StatefulWidget {
