+++ adaptive_navigation/example/lib/default_scaffold.dart
@@ -1,4 +1,4 @@
-import 'package:adaptive_navigation/scaffold.dart';
+import 'package:adaptive_navigation/adaptive_navigation.dart';
 import 'package:flutter/material.dart';
 
 void main() {
