+++ adaptive_navigation/lib/adaptive_navigation.dart
@@ -1,0 +1,8 @@
+// Copyright 2020, the Flutter project authors. Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+library adaptive_navigation;
+
+export 'src/adaptive_app_bar.dart';
+export 'src/adaptive_navigation_scaffold.dart';
