+++ adaptive_navigation/lib/src/adaptive_app_bar.dart
@@ -1,0 +1,65 @@
+// Copyright 2020, the Flutter project authors. Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'package:flutter/material.dart';
+
+/// AdaptiveAppBar has a leading width of 72.0. Everything else is the same as
+/// [AppBar].
+class AdaptiveAppBar extends AppBar {
+  AdaptiveAppBar({
+    Key key,
+    Widget leading,
+    bool automaticallyImplyLeading = true,
+    Widget title,
+    List<Widget> actions,
+    Widget flexibleSpace,
+    PreferredSizeWidget bottom,
+    double elevation,
+    Color shadowColor,
+    ShapeBorder shape,
+    Color backgroundColor,
+    Brightness brightness,
+    IconThemeData iconThemeData,
+    IconThemeData actionsIconThemeData,
+    TextTheme textTheme,
+    bool primary = true,
+    bool centerTitle,
+    bool excludeHeaderSemantics = false,
+    double titleSpacing = NavigationToolbar.kMiddleSpacing,
+    double toolbarOpacity = 1.0,
+    double bottomOpacity = 1.0,
+    double toolbarHeight,
+    double leadingWidth = 72.0,
+  })  : assert(automaticallyImplyLeading != null),
+        assert(elevation == null || elevation >= 0.0),
+        assert(primary != null),
+        assert(titleSpacing != null),
+        assert(toolbarOpacity != null),
+        assert(bottomOpacity != null),
+        super(
+        key: key,
+        leading: leading,
+        automaticallyImplyLeading: automaticallyImplyLeading,
+        title: title,
+        actions: actions,
+        flexibleSpace: flexibleSpace,
+        bottom: bottom,
+        elevation: elevation,
+        shadowColor: shadowColor,
+        shape: shape,
+        backgroundColor: backgroundColor,
+        brightness: brightness,
+        iconTheme: iconThemeData,
+        actionsIconTheme: actionsIconThemeData,
+        textTheme: textTheme,
+        primary: primary,
+        centerTitle: centerTitle,
+        excludeHeaderSemantics: excludeHeaderSemantics,
+        titleSpacing: titleSpacing,
+        toolbarOpacity: toolbarOpacity,
+        bottomOpacity: bottomOpacity,
+        toolbarHeight: toolbarHeight,
+        leadingWidth: leadingWidth,
+      );
+}
