+++ adaptive_navigation/lib/src/adaptive_navigation_scaffold.dart
@@ -207,7 +207,7 @@ class AdaptiveNavigationScaffold extends StatelessWidget {
     );
     final drawerDestinations = destinations.length > bottomNavigationOverflow
         ? destinations.sublist(
-            includeBaseDestinationsInMenu ? 0 : bottomNavigationOverflow)
+        includeBaseDestinationsInMenu ? 0 : bottomNavigationOverflow)
         : [];
     return Scaffold(
       body: body,
@@ -239,7 +239,7 @@ class AdaptiveNavigationScaffold extends StatelessWidget {
     );
     final drawerDestinations = destinations.length > railDestinationsOverflow
         ? destinations.sublist(
-            includeBaseDestinationsInMenu ? 0 : railDestinationsOverflow)
+        includeBaseDestinationsInMenu ? 0 : railDestinationsOverflow)
         : [];
     return Scaffold(
       appBar: appBar,
@@ -396,66 +396,6 @@ class AdaptiveNavigationScaffold extends StatelessWidget {
       onDestinationSelected(index);
     }
   }
-}
-
-/// AdaptiveAppBar has a leading width of 72.0. Everything else is the same as
-/// [AppBar].
-class AdaptiveAppBar extends AppBar {
-  AdaptiveAppBar({
-    Key key,
-    Widget leading,
-    bool automaticallyImplyLeading = true,
-    Widget title,
-    List<Widget> actions,
-    Widget flexibleSpace,
-    PreferredSizeWidget bottom,
-    double elevation,
-    Color shadowColor,
-    ShapeBorder shape,
-    Color backgroundColor,
-    Brightness brightness,
-    IconThemeData iconThemeData,
-    IconThemeData actionsIconThemeData,
-    TextTheme textTheme,
-    bool primary = true,
-    bool centerTitle,
-    bool excludeHeaderSemantics = false,
-    double titleSpacing = NavigationToolbar.kMiddleSpacing,
-    double toolbarOpacity = 1.0,
-    double bottomOpacity = 1.0,
-    double toolbarHeight,
-    double leadingWidth = 72.0,
-  })  : assert(automaticallyImplyLeading != null),
-        assert(elevation == null || elevation >= 0.0),
-        assert(primary != null),
-        assert(titleSpacing != null),
-        assert(toolbarOpacity != null),
-        assert(bottomOpacity != null),
-        super(
-          key: key,
-          leading: leading,
-          automaticallyImplyLeading: automaticallyImplyLeading,
-          title: title,
-          actions: actions,
-          flexibleSpace: flexibleSpace,
-          bottom: bottom,
-          elevation: elevation,
-          shadowColor: shadowColor,
-          shape: shape,
-          backgroundColor: backgroundColor,
-          brightness: brightness,
-          iconTheme: iconThemeData,
-          actionsIconTheme: actionsIconThemeData,
-          textTheme: textTheme,
-          primary: primary,
-          centerTitle: centerTitle,
-          excludeHeaderSemantics: excludeHeaderSemantics,
-          titleSpacing: titleSpacing,
-          toolbarOpacity: toolbarOpacity,
-          bottomOpacity: bottomOpacity,
-          toolbarHeight: toolbarHeight,
-          leadingWidth: leadingWidth,
-        );
 }
 
 bool _isLargeScreen(BuildContext context) =>
