+++ adaptive_navigation/pubspec.yaml
@@ -1,7 +1,7 @@
 name: adaptive_navigation
 description: A component to be used in place of the Scaffold widget when you want automatic switching between Navigation Rail and Bottom Navigation Bar.
 version: 0.0.1
-homepage: TBD
+homepage: https://github.com/material-components/material-components-flutter-adaptive
 
 environment:
   sdk: ">=2.7.0 <3.0.0"
@@ -9,9 +9,9 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  adaptive_breakpoints: ^0.0.2
+#    path: ../adaptive_breakpoints
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  adaptive_breakpoints:
-    path: ../adaptive_breakpoints
